Moved homepage to test for and pointed clone to that file
