+++ frontend/src/components/factories/mainViewFactory.js
@@ -3,6 +3,8 @@ const edaMainViewHandler = require(`../modules/eda/edaMainViewHandler`);
 const policyMainViewHandler = require(`../modules/policy/policyMainViewHandler`);
 const globalSearchMainViewHandler = require(`../modules/globalSearch/globalSearchMainViewHandler`);
 
+const policyTestMainViewHandler = require(`../modules/policy/policyTestMainViewHandler`)
+
 class MainViewFactory {
 
 	constructor(module) {
@@ -10,6 +12,10 @@ class MainViewFactory {
 			switch (module) {
 				case 'policy/policyMainViewHandler':
 					this.handler = policyMainViewHandler;
+					break;
+				// TODO Remove this once the homepage is used for main GC page. 
+				case 'policy/policyTestMainViewHandler':
+					this.handler = policyTestMainViewHandler;
 					break;
 				case 'eda/edaMainViewHandler':
 					this.handler = edaMainViewHandler;
